implement cluster as Class refactoring / renaming for better readability
